ClientHolder catch exception during shutdown.
When running shutdown on the client poolers, some client may throw
exceptions that causes a part of the shutdown process not to be
launched. This may cause database inconsistencies if structure are not
shut down properly.
The added code, run shutdown for all client and throw an exception at
the end if any.