revert Travis & Composer mess
In order to allow HHVM test suite to run on Travis, it has been set to
ignore plate-form requirements. This lead Atoum to install the 3.x
version of the test suite which is incompatible with PHP 5.5.

It has been decided to ignore HHVM on Travis. Hopefully people using
HHVM are used to test the software they use on this plate-form.